+++ macros/CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add derive macros for the `Serialize`, `Deserialize` and `Introspectable` traits.
 - Add derive macros for the `SerializeKey`, `DeserializeKey` and `KeyTypeOf` traits.
+- Add a derive macro for the `AsSerializeArg` trait.
 
 ### Changed